AI_Workspace/backend/requirements.txt
# ============================================================
# Combined Requirements — Botivate AI Workspace
# ============================================================

# --- Core Framework ---
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.0

# --- HTTP Client ---
httpx>=0.28.0

# --- Database ---
sqlalchemy>=2.0.0
aiosqlite>=0.20.0

# --- Authentication ---
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9

# --- LLM Providers ---
openai>=1.50.0
groq>=0.11.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langgraph>=0.2.0

# --- Vector Store ---
chromadb>=0.5.0
sentence-transformers>=3.0.0

# --- Google APIs ---
google-api-python-client>=2.140.0
google-auth>=2.34.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
gspread>=6.1.0

# --- Email ---
aiosmtplib>=3.0.0

# --- Document Parsing ---
PyPDF2>=3.0.0
pypdf>=4.0.0
python-docx>=1.1.0
pdfplumber>=0.11.0

# --- NLP ---
spacy>=3.7.0

# --- ML/AI ---
transformers>=4.40.0
torch>=2.0.0

# --- Scheduling ---
apscheduler>=3.10.0

# --- Utilities ---
jinja2>=3.1.0
aiofiles>=24.1.0
